+++ core-service/src/main/java/com/microservice/coreservice/domain/dto/TaskDto.java
@@ -13,15 +13,15 @@ import java.sql.Timestamp;
 public class TaskDto {
 
     private String name;
-    private int assignee;
+    private Long assignee;
     private Long startDay;
     private Long endDay;
     private String priority;
     private String jobDescription;
     private String status;
-    private int parentId;
-    private Integer projectId;
-    private Integer sectionId;
+    private Long parentId;
+    private Long projectId;
+    private Long sectionId;
     private float estimate_time;
     private Long created_date;
     private Long updated_date;
