+++ core-service/src/main/java/com/microservice/coreservice/repository/ProjectRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ProjectRepository extends CrudRepository<Project, Integer> {
+public interface ProjectRepository extends CrudRepository<Project, Long> {
     boolean existsByName(String name);
 
 }
