+++ core-service/src/main/java/com/microservice/coreservice/repository/ProjectTeamRepository.java
@@ -7,13 +7,13 @@ import org.springframework.data.repository.CrudRepository;
 
 import javax.transaction.Transactional;
 
-public interface ProjectTeamRepository extends CrudRepository<ProjectTeam, Integer> {
+public interface ProjectTeamRepository extends CrudRepository<ProjectTeam, Long> {
     @Query(value = "SELECT new com.microservice.coreservice.entity.ProjectTeam (p.project_id, p.team_id, p.created_date, p.updated_date)  FROM ProjectTeam p WHERE p.team_id = ?1")
-    Object findByTeamId(int teamId);
+    Object findByTeamId(Long teamId);
 
     @Transactional
     @Modifying
     @Query(value = "DELETE ProjectTeam p WHERE p.team_id = :id")
-    void deleteByTeamId(int id);
+    void deleteByTeamId(Long id);
 
 }
