+++ core-service/src/main/java/com/microservice/coreservice/repository/SectionRepository.java
@@ -3,5 +3,5 @@ package com.microservice.coreservice.repository;
 import com.microservice.coreservice.entity.Section;
 import org.springframework.data.repository.CrudRepository;
 
-public interface SectionRepository extends CrudRepository<Section, Integer> {
+public interface SectionRepository extends CrudRepository<Section, Long> {
 }
