+++ core-service/src/main/java/com/microservice/coreservice/repository/TeamRepository.java
@@ -6,11 +6,11 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface TeamRepository extends CrudRepository<Team, Integer> {
+public interface TeamRepository extends CrudRepository<Team, Long> {
     boolean existsByName(String name);
 
     @Query("SELECT new com.microservice.coreservice.entity.Team(t.id, t.name, t.description, t.enabled, t.created_date, t.updated_date) FROM Team t JOIN ProjectTeam pt on t.id = pt.team_id WHERE pt.project_id = :projectId")
-    List<Team> findByProjectId(int projectId);
+    List<Team> findByProjectId(Long projectId);
 
 
 }
